pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stem-separator"
version = "2.0.0"
description = "Separate audio into stems using AI (Demucs)"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Stem Separator Contributors"}
]
keywords = [
    "audio",
    "music",
    "stems",
    "separation",
    "demucs",
    "vocals",
    "drums",
    "bass",
    "karaoke",
    "acapella",
    "ai",
    "machine-learning",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Multimedia :: Sound/Audio :: Analysis",
    "Topic :: Multimedia :: Sound/Audio :: Conversion",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.9"
dependencies = [
    "demucs>=4.0.0",
    "yt-dlp>=2023.1.1",
    "soundfile>=0.12.0",
    "scipy>=1.9.0",
    "numpy>=1.21.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
spotify = ["spotdl>=4.0.0"]
preview = ["sounddevice>=0.4.0"]
metadata = ["mutagen>=1.45.0"]
gpu = ["torch>=2.0.0", "torchaudio>=2.0.0"]
all = [
    "spotdl>=4.0.0",
    "sounddevice>=0.4.0",
    "mutagen>=1.45.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/stem-separator/stem-separator-cli"
Documentation = "https://github.com/stem-separator/stem-separator-cli#readme"
Repository = "https://github.com/stem-separator/stem-separator-cli"
Issues = "https://github.com/stem-separator/stem-separator-cli/issues"

[project.scripts]
stem-separator = "stem_separator.cli:main"
stems = "stem_separator.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["stem_separator*"]

[tool.setuptools.package-data]
stem_separator = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "S",    # flake8-bandit (security)
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # function call in default argument
    "S101",   # use of assert
    "S603",   # subprocess without shell=True (we need this)
    "S607",   # partial executable path
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=stem_separator --cov-report=term-missing"
